Complete spectrum annotation and quant mad
In order to better localise the modifications, we
need to get the theoretical assignment of fragments
to the spectrum. From this, we should be able
to make a ladder, and find out how specific the mod
assignment is.

For cases where we have ambiguous mods, the quants
may include distributions of more than one peptide
species (representing the quants for the different
sets of mods). We note down the MAD (median absolute
deviation) so that we can possibly detect if there
is more than one distribution.